+++ backend/public_app/service-worker.js
@@ -1,4 +1,4 @@
-const RELEASE="3.23.7.1";
+const RELEASE="3.23.7.2";
 const CACHE_PREFIX="scsi-";
 const VERSION=`${CACHE_PREFIX}v${RELEASE}`;
 const IMMUTABLE=`${VERSION}-immutable`;